修改监视器

+++ watch.py
@@ -7,6 +7,8 @@ import time
 from watchdog.events import FileSystemEventHandler
 from watchdog.observers import Observer
 
+from common import console
+
 control_p = None
 p = None
 
@@ -14,7 +16,12 @@ p = None
 class when_file_chanage(FileSystemEventHandler):
     def on_any_event(self, event):
         if not event.is_directory and not event.src_path.endswith(('.pyc', '.swp', ".swx")):
-            p.send_signal(1)
+            if os.path.basename(os.path.dirname(event.src_path)) != "static_page":
+                console.log("info", "File [{}] detected a change, the program is restarting".format(event.src_path))
+                p.send_signal(1)
+            if os.path.basename(event.src_path) == "control_server.py":
+                if control_p is not None:
+                    control_p.send_signal(1)
 
 
 def HUP_handler(signum, frame):
